Utility to promote file extensions.
If there are files in a directory that ends with a specific extension,
but the extension is not preceded by a period, such files are renamed
so that their new filename has the extension separated separated from
the filename by a perion. E. g. examplehtml is renamed to example.html